backend/requirements.txt
autopep8==2.0.2
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
dnspython==2.3.0
ExifRead==3.0.0
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pefile==2023.2.7
protobuf==4.21.12
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.14.0
pymongo==4.3.3
python-dotenv==1.0.0
requests==2.30.0
urllib3==2.0.2
volatility3==2.4.2
Werkzeug==2.3.4
